Use {client_name} if available instead of hostname passed to xxfi_connect
Spamassassin's RDNS_NONE check tests if the rdns field in the Received
header is empty or "unknown". Postfix sets the client_name macro to
"unknown" when the FcrDNS check fails, but the hostname passed to
xxfi_connect is set to the IP address which means that RDNS_NONE will
never trigger. Fix this by using client_name if available and only fall
back to the hostname parameter otherwise.

Signed-off-by: Florian Pritz <bluewind@xinu.at>