Use the {client_name} macro if available instead of the hostname as RDNS in Received header
The Spamassassin's RDNS_NONE check tests if the rdns field in
the Received header is empty or "unknown".  Postfix sets
the {client_name} macro to "unknown" when the FcrDNS check
fails, but the hostname passed to xxfi_connect is set to
the IP address which means that RDNS_NONE is never triggered.
Instead, use {client_name} macro, and only if it is not available
use the hostname.

Signed-off-by: Florian Pritz <bluewind@xinu.at>
Closes: #1, #3