Refactor mlfi_connect client_name handling
Spamassassin's RDNS_NONE check requires in the synthesized Received: header verified reverse DNS hostname or "unknown" if the reverse DNS is not available.

Sendmail sends an IP address in square brackets in the client_host argument and in the {client_name} macro, if the verified reverse DNS is not available

Postfix sends a verified reverse DNS or "unknown" using the {client_name} macro.

Signed-off-by: Petr Řehoř <rx@rx.cz>